# apps/go-api/internal/types/openapi.gen.go
// Package types provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package types

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	RoleScopes = "Role.Scopes"
)

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	// ContactEmail 連絡用メールアドレス
	ContactEmail *openapi_types.Email `json:"contactEmail"`

	// FamilyName 姓
	FamilyName string `json:"familyName"`

	// GivenName 名
	GivenName string `json:"givenName"`

	// PhoneNumber 電話番号
	PhoneNumber *string `json:"phoneNumber"`
}

// Error defines model for Error.
type Error struct {
	// Code エラーコード
	Code int `json:"code"`

	// Message エラーメッセージ
	Message string `json:"message"`
}

// UpdateUserRequest defines model for UpdateUserRequest.
type UpdateUserRequest struct {
	// ContactEmail 連絡用メールアドレス
	ContactEmail *openapi_types.Email `json:"contactEmail"`

	// FamilyName 姓
	FamilyName string `json:"familyName"`

	// GivenName 名
	GivenName string `json:"givenName"`

	// PhoneNumber 電話番号
	PhoneNumber *string `json:"phoneNumber"`
}

// User defines model for User.
type User struct {
	// ContactEmail 連絡用メールアドレス
	ContactEmail *openapi_types.Email `json:"contactEmail"`

	// CreatedAt 作成日時
	CreatedAt time.Time `json:"createdAt"`

	// Email メールアドレス（理科大メール）
	Email openapi_types.Email `json:"email"`

	// FamilyName 姓
	FamilyName string `json:"familyName"`

	// GivenName 名
	GivenName string `json:"givenName"`

	// Id ユーザーID
	Id int `json:"id"`

	// IsActive アカウント有効フラグ
	IsActive bool `json:"isActive"`

	// IsVerified メール認証済み
	IsVerified bool `json:"isVerified"`

	// PhoneNumber 電話番号
	PhoneNumber *string `json:"phoneNumber"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updatedAt"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = CreateUserRequest

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = UpdateUserRequest